Logging makes everything better
and check if media file before reading exif data

+++ src/lib.rs
@@ -1,3 +1,6 @@
+#[macro_use]
+extern crate log;
+
 use clap::Parser;
 use exif::{In, Tag};
 use lazy_static::lazy_static;
@@ -37,9 +40,16 @@ impl MediaConfig {
 
     pub fn copy_media_files(&mut self) -> Result<(), Box<dyn Error>> {
         self.find_all_media_files(None, true)?;
+        info!("Found {} files", self.files.len());
+        let mut copied_files = 0;
         for (source, target) in self.files.iter() {
-            copy_file(source, target)?;
+            match copy_file(source, target) {
+                Ok(true) => copied_files += 1,
+                Ok(false) => (),
+                Err(e) => error!("Error copying file: {}", e),
+            }
         }
+        info!("Copied {}/{} files", copied_files, self.files.len());
         Ok(())
     }
 
@@ -54,7 +64,7 @@ impl MediaConfig {
             let path = entry.path();
             if path.is_dir() && recursive {
                 self.find_all_media_files(Some(path.to_str().unwrap()), true)?;
-            } else if path.is_file() {
+            } else if path.is_file() && is_media_file(&path) {
                 let sourcepath = &path.to_str().unwrap();
                 if let Some(targetpath) = smartphone_file(sourcepath) {
                     self.files
@@ -69,6 +79,20 @@ impl MediaConfig {
     }
 }
 
+fn is_media_file(path: &Path) -> bool {
+    let ext = path.extension();
+
+    match ext {
+        None => false,
+        Some(file_ext) => matches!(
+            String::from(file_ext.to_str().unwrap())
+                .to_lowercase()
+                .as_str(),
+            "jpg" | "jpeg" | "mp4" | "png"
+        ),
+    }
+}
+
 pub fn run(args: Args) -> Result<(), Box<dyn Error>> {
     let home = env::var("HOME")?;
     let target = Path::new(&home).join("Pictures");
@@ -77,13 +101,17 @@ pub fn run(args: Args) -> Result<(), Box<dyn Error>> {
 }
 
 // Copy file from one directory to another
-fn copy_file(from: &str, to: &str) -> Result<(), Box<dyn Error>> {
+fn copy_file(from: &str, to: &str) -> Result<bool, Box<dyn Error>> {
     let abs_path = Path::new(&to);
     let parent = abs_path.parent().unwrap();
     create_dir(parent.to_str().unwrap())?;
-    println!("Copy file {} to {}", from, abs_path.to_str().unwrap());
+    if abs_path.exists() {
+        warn!("Skipping File {}, already exists", to);
+        return Ok(false);
+    }
+    info!("Copy file {} to {}", from, abs_path.to_str().unwrap());
     fs::copy(from, to)?;
-    Ok(())
+    Ok(true)
 }
 
 // Create directory, if it does not exist
@@ -107,41 +135,39 @@ fn smartphone_file(filename: &str) -> Option<String> {
         .unwrap();
     };
 
-    RE.captures(filename).and_then(|cap| {
-        Some(format!(
-            "{}/{}/{}/{}",
-            &cap["y"], &cap["m"], &cap["d"], &cap[0]
-        ))
-    })
+    RE.captures(filename)
+        .map(|cap| format!("{}/{}/{}/{}", &cap["y"], &cap["m"], &cap["d"], &cap[0]))
 }
 
 fn read_jpg_exif(filename: &str) -> Option<String> {
+    // filename needs to end with .jpg or .png
+    if !filename.to_lowercase().ends_with(".jpg") && !filename.to_lowercase().ends_with(".png") {
+        return None;
+    }
     lazy_static! {
         static ref RE: Regex =
             Regex::new(r"(?P<y>\d{4})-(?P<m>\d{2})-(?P<d>\d{2})\s+(?:\d|:){8}").unwrap();
     };
-    let file = File::open(filename).expect(format!("Could not open file {}", filename).as_str());
+    let file = File::open(filename).unwrap_or_else(|_| panic!("Could not open file {}", filename));
     let mut bufreader = std::io::BufReader::new(&file);
     let exifreader = exif::Reader::new();
     let exif = exifreader.read_from_container(&mut bufreader).unwrap();
     let datetime = match exif.get_field(Tag::DateTimeOriginal, In::PRIMARY) {
-        Some(field) => {
-            println!("{}: {}", field.tag, field.display_value());
-            RE.captures(field.display_value().to_string().as_str())
-                .and_then(|cap| {
-                    Some(format!(
-                        "{}/{}/{}/{}",
-                        &cap["y"],
-                        &cap["m"],
-                        &cap["d"],
-                        Path::new(filename)
-                            .file_name()
-                            .expect("no filename")
-                            .to_str()
-                            .unwrap()
-                    ))
-                })
-        }
+        Some(field) => RE
+            .captures(field.display_value().to_string().as_str())
+            .map(|cap| {
+                format!(
+                    "{}/{}/{}/{}",
+                    &cap["y"],
+                    &cap["m"],
+                    &cap["d"],
+                    Path::new(filename)
+                        .file_name()
+                        .expect("no filename")
+                        .to_str()
+                        .unwrap()
+                )
+            }),
         _ => Some(String::from("no exif data")),
     };
     datetime
@@ -157,6 +183,20 @@ mod tests {
         ($fname:expr) => {
             concat!(env!("CARGO_MANIFEST_DIR"), "/tests/", $fname)
         };
+    }
+
+    #[test]
+    fn test_is_media_file() {
+        let list_of_media_files = vec!["jpg", "jpeg", "mp4", "png", "JPG", "JPEG", "MP4", "PNG"];
+        for media_file in list_of_media_files {
+            let filename = format!("test.{}", media_file);
+            assert_eq!(
+                true,
+                is_media_file(Path::new(&filename)),
+                "File should be a media file {}",
+                filename
+            );
+        }
     }
 
     #[test]
